# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nonce Auth API
  description: Open-source Authentication Backend-as-a-Service (Auth BaaS). Multi-tenant, project-scoped auth.
  version: 0.1.0
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080/v1
    description: Local (API base path /v1)
paths:
...
  /auth/refresh:
    post:
      summary: Refresh tokens
      description: Exchange a valid refresh token for a new access token and optionally a new refresh token.
      operationId: refresh
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /users/me:
    get:
      summary: Get current user
      description: Returns the authenticated user (from JWT). Requires JWT.
      operationId: getMe
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User not found
    patch:
      summary: Update current user
      description: Update current user profile (e.g. email, user_metadata). Requires JWT.
      operationId: updateMe
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                user_metadata:
                  type: object
      responses:
        "200":
          description: Updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Delete current user (GDPR)
      description: Soft-delete the current user and revoke all sessions. Requires JWT.
      operationId: deleteMe
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        "204":
          description: User deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/export:
    get:
      summary: Export current user data (GDPR)
      description: Returns the current user's data for data portability. Requires JWT.
      operationId: exportMe
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User data export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportMeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    ProjectKey:
      type: apiKey
      in: header
      name: X-Nonce-Project-Key
      description: Project API key (or Authorization Bearer with project key)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for protected routes
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          maxLength: 1024
...
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
...
    MeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        anonymous:
          type: boolean
        user_metadata:
          type: object
        app_metadata:
          type: object
    ExportMeResponse:
      type: object
      properties:
        exported_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/MeResponse"
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request body or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid project key / credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"